requirements.txt
streamlit
moviepy
openai-whisper
transformers
torch
spacy
en-core-web-sm